chore: 코드 폴더 위치 수정
