+++ FlatBread/Features/Payment/Presentation/Model/IamportPaymentInput.swift
@@ -1,0 +1,15 @@
+//
+//  IamportPaymentInput.swift
+//  FlatBread
+//
+//  Created by andev on 11/25/25.
+//
+
+import Foundation
+
+struct IamportPaymentInput {
+    let postId: String
+    let price: Int
+    let title: String
+    let buyerName: String
+}
